refactor(al4): isolate send write context
Finding: RULE-003
